Refactoring Expense Groups
Way too many things going on in expense groups.

One such item of note is that I am including an Options Dialog with every expense group.

If we start getting a lot of groups, we may effect performance. So, I'm moving that dialog up into the App component for readability and ease of use reasons.